Create new Index and separate useMachine and useService
